Mongo_learning_app/requirements.txt
pymongo == --upgrade pymongo
python-dotenv==0.21.0
rich==13.3.1

"""
    Para actualizar dependencias ejecutar en terminal el comando:

        pip install --upgrade -r requirements.txt

"""